pyproject.toml
[project]
name = "colgate-kb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "gradio>=6.13.0",
    "langchain>=1.2.15",
    "langchain-groq>=1.1.2",
    "langchain-ollama>=1.1.0",
    "lxml>=6.1.0",
    "pdfplumber>=0.11.9",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "selenium>=4.43.0",
    "trafilatura>=2.0.0",
    "webdriver-manager>=4.0.2",
    "yt-dlp>=2026.3.17",
]
